requirements.txt
Django>=5.0,<6.0
gunicorn>=21.2.0
whitenoise>=6.5.0
psycopg2-binary>=2.9.9
django-unfold>=0.40.0
stripe>=7.0.0
Pillow>=10.0.0
dnspython>=2.6.0
beautifulsoup4>=4.12.0
reportlab>=4.0.0
requests>=2.31.0
